DF-6084 fix review comments. set warnings as errors on ios, android, win32
Former-commit-id: e013e6bd35bce1defa328d42e83669b26b4d4626

+++ Sources/Internal/Utils/UTF8UtilsAndroid.cpp
@@ -39,7 +39,7 @@
 namespace DAVA 
 {
 
-void UTF8Utils::EncodeToWideString(const uint8 * string, int32 size, WideString & resultString)
+void UTF8Utils::EncodeToWideString(const uint8 * string, size_t size, WideString & resultString)
 {
 	resultString = L"";
